Big change in Variant and String, moving size_t, int and uint in serperate set/append methods

+++ CcNetwork/CcHttp/CcHttpRequest.cpp
@@ -209,7 +209,7 @@ void CcHttpRequest::setContentType(const CcString& sContentType)
   appendHeaderLine(CcHttpGlobalStrings::Header::ContentType, sContentType);
 }
 
-void CcHttpRequest::setContentLength(size_t uiSize)
+void CcHttpRequest::setContentLength(uint64 uiSize)
 {
   appendHeaderLine(CcHttpGlobalStrings::Header::ContentLength, CcString::fromNumber(uiSize));
 }
